+++ wemake_python_styleguide/logic/naming/builtins.py
@@ -3,13 +3,10 @@ import inspect
 import keyword
 from typing import Final
 
-from wemake_python_styleguide.constants import BUILTINS_WHITELIST
 from wemake_python_styleguide.logic.naming.access import is_magic, is_unused
 
 _BUILTINS: Final = frozenset(
-    builtin[0]
-    for builtin in inspect.getmembers(builtins)
-    if builtin[0] not in BUILTINS_WHITELIST
+    builtin[0] for builtin in inspect.getmembers(builtins)
 )
 
 _ALL_BUILTINS: Final = frozenset(